refactor: fish canvas to standalone component
